Original commit eeeee1d77eb5d17685c4f4ddfb8107daef732131

Some weird stuff with boolean vs. dboolean

Mac still builds and runs

+++ src/engine/i_audio.c
@@ -173,8 +173,8 @@ typedef struct {
     dword       nexttic;
     dword       lasttic;
     dword       starttic;
-    uint32_t      starttime;
-    uint32_t      curtime;
+    uint64_t      starttime;
+    uint64_t      curtime;
     chanstate_e state;
     dboolean    paused;
